ref: Moved all metatable related custom functions to its own file
