+++ CMakeLists.txt
@@ -38,6 +38,8 @@ add_library(Module SHARED
         ClosureManager.hpp
         Environment/Libraries/Closures.cpp
         Environment/Libraries/Closures.hpp
+        Environment/Libraries/Metatable.cpp
+        Environment/Libraries/Metatable.hpp
 )
 target_include_directories(Module PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}")
 target_include_directories(Module PRIVATE "${CMAKE_CURRENT_SOURCE_DIR}/Dependencies")
